+++ Assets/Objects/Abilities/LightningStrike/LightningStrikeProjectile.cs
@@ -29,7 +29,7 @@ namespace Objects.Abilities.LightningStrike
 			lightingChainProjectile.SetStats(new WeaponStats()
 			{
 				TimeToLive = 0.5f,
-				Damage = (int)(WeaponStats.GetDamage() * 0.25f),
+				Damage = WeaponStats.GetDamage() * 0.25f,
 				Scale = WeaponStats.GetScale(),
 				DetectionRange = 1f
 			});
